fixup provisioner
